v0.24

+++ src/main/java/com/javarush/jira/bugtracking/task/TaskTagService.java
@@ -1,5 +1,7 @@
 package com.javarush.jira.bugtracking.task;
 
+import com.javarush.jira.common.error.DataConflictException;
+import com.javarush.jira.login.AuthUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,18 +13,21 @@ public class TaskTagService {
 
     private final TaskTagRepository repository;
 
-    public List<Task> getAllTasksForTagNames(List<String> tags) {
-        List<Task> tasks = repository.findAllTasksByTagNames(tags);
+    public List<Long> getAllTasksForTagNames(List<String> tags) {
+        List<Long> tasks = repository.findAllTasksByTagNames(tags);
         return tasks;
     }
 
-    public List<TaskTag> getAllTagsForTask(long id) {
-        List<TaskTag> taskTags = repository.findAllTagsForTask(id);
+    public List<String> getAllTagsForTask(long id) {
+        List<String> taskTags = repository.findAllTagsForTask(id);
         return taskTags;
     }
 
     public void updateTagsForTask(long id, List<String> taskTags){
-        List<TaskTag> oldTaskTags = repository.findAllTagsForTask(id);
+        List<TaskTag> oldTaskTags = repository.findAllTaskTagsForTask(id);
+        if (oldTaskTags.size() == 0) {
+            throw new DataConflictException("Task " + id + " doesn't exist.");
+        }
         for (TaskTag taskTag:
             oldTaskTags) {
             repository.delete(taskTag);
